requirements.txt
openai
opencv-python
numpy
requests
paddleocr
paddlepaddle
websocket-client
python-dotenv
apscheduler